+++ test/cider_ns_tests.clj
@@ -1,0 +1,2 @@
+(ns cider-ns-tests
+  "Supports cider-ns-tests.el")